Rename: 대시보드 타입 파일 이름 수정
