+++ src/components/Dashboard/DashboardItem.tsx
@@ -1,6 +1,6 @@
 import { NavLink } from 'react-router-dom';
 import crown from 'src/assets/images/crown.svg';
-import { Dashboard } from '../../types/api';
+import { Dashboard } from '../../types/dashboardTypes';
 
 export default function DashboardItem({ dashboard }: { dashboard: Dashboard }) {
   const colorVariants: Record<string, string> = {
